Fixed source setup for "." package path
On clone URLs which directly points to the source
the package path is the current directory relative
to the clone path. In this case we need to create
a package subdir to allow the tar process to
continue to work. Example:

corbos_scm --git https://git.launchpad.net/ubuntu/+source/curl \
    --package . --outdir ~/