Refactor to use ubuntu tools
All tools needed are available in the ubuntu-dev-tools
package. However, these tools are supposed to run in a
Debian environment. Because of this the obs service fetches
an application container providing ubuntu-dev-tools and
calls the relevant fetcher through podman. This allows
to make use of service in obs independently of the distribution
used for the OBS server. This Fixes #2