Fixup volume name limitations
Create a safe symlink with regards to the volume
name limitations of podman. For sharing a directory
some characters are not allowed, e.g ":". In OBS
the colon is used as project separator. Because of
that it happens very easily that the --outdir path
created by OBS at call time of the service contains
colons and prevents this directory name from being
eligible to be shared via --volume with podman.
The recommended workaround from the podman team is
to create a temporary symlink